some javadoc

+++ src/main/kotlin/io/craftray/huntrace/game/GameWorldController.kt
@@ -8,7 +8,19 @@ import org.bukkit.World
 
 class GameWorldController(private val game: Game) {
 
+    private var generated = false
+    private var linked = false
+    private var unlink = false
+    private var deleted = false
+
+    /**
+     * Generate the three dimensions of a game
+     * @author Kylepoops
+     * @return the world set
+     */
     fun generateWorlds(): WorldSet {
+        if (generated)  throw IllegalStateException("Worlds have already been generated")
+
         val rule = game.rules.worldRule
         val worlds = WorldSet()
 
@@ -34,6 +46,8 @@ class GameWorldController(private val game: Game) {
             seed = rule.seed
         }.build()
 
+        this.generated = true
+
         return worlds.apply {
             this.overworld = overworld
             this.nether = nether
@@ -41,23 +55,46 @@ class GameWorldController(private val game: Game) {
         }
     }
 
+    /**
+     * Delete the worlds of the game
+     * @author Kylepoops
+     * @exception IllegalStateException if the worlds have not been generated or have been deleted
+     */
     fun deleteWorlds() = game.worlds.run {
+        if (!generated) throw IllegalStateException("Worlds have not been generated")
+        if (deleted) throw IllegalStateException("Worlds have already been deleted")
         this.overworld.delete()
         this.nether.delete()
         this.theEnd.delete()
+        deleted = true
     }
 
-    fun linkWorlds() {
-        game.worlds.run {
-            MultiverseWorldManager.linkWorlds(this.overworld, this.nether, this.theEnd)
-            MultiverseWorldManager.linkInventories(this.overworld, this.nether, this.theEnd)
-        }
+    /**
+     * Link the three dimensions
+     * @author Kylepoops
+     * @exception IllegalStateException if the worlds have not been generated or have been deleted or linked
+     */
+    fun linkWorlds() = game.worlds.run {
+        if (!generated) throw IllegalStateException("Worlds have not been generated")
+        if (deleted) throw IllegalStateException("Worlds have already been deleted")
+        if (linked) throw IllegalStateException("Worlds have already been linked")
+        MultiverseWorldManager.linkWorlds(this.overworld, this.nether, this.theEnd)
+        MultiverseWorldManager.linkInventories(this.overworld, this.nether, this.theEnd)
+        linked = true
     }
 
-    fun unlinkWorlds() {
-        game.worlds.run {
-            MultiverseWorldManager.unlinkWorlds(this.overworld, this.nether, this.theEnd)
-            MultiverseWorldManager.unlinkInventories(this.overworld, this.nether, this.theEnd)
-        }
+    /**
+     * Unlink the three dimensions
+     * @author Kylepoops
+     * @exception IllegalStateException if the worlds have not been generated or have been deleted, linked or unlinked
+     */
+    fun unlinkWorlds() = game.worlds.run {
+        if(!generated) throw IllegalStateException("Worlds have not been generated")
+        if(deleted) throw IllegalStateException("Worlds have already been deleted")
+        if (!linked) throw IllegalStateException("Worlds have not been linked")
+        if (unlink) throw IllegalStateException("Worlds have already been unlinked")
+        MultiverseWorldManager.unlinkWorlds(this.overworld, this.nether, this.theEnd)
+        MultiverseWorldManager.unlinkInventories(this.overworld, this.nether, this.theEnd)
+        unlink = false
     }
 }
